+++ apps/cowswap-frontend/src/common/pure/CurrencyArrowSeparator/index.tsx
@@ -15,15 +15,14 @@ export interface CurrencyArrowSeparatorProps {
   onSwitchTokens(): void
 }
 
-export function CurrencyArrowSeparator(props: CurrencyArrowSeparatorProps): ReactNode {
-  const {
-    isLoading,
-    onSwitchTokens,
-    isCollapsed = true,
-    hasSeparatorLine,
-    disabled = false,
-    isDarkMode = false,
-  } = props
+export function CurrencyArrowSeparator({
+  isLoading,
+  onSwitchTokens,
+  isCollapsed = true,
+  hasSeparatorLine,
+  disabled = false,
+  isDarkMode = false,
+}: CurrencyArrowSeparatorProps): ReactNode {
   const isInjectedWidgetMode = isInjectedWidget()
 
   return (
@@ -32,9 +31,8 @@ export function CurrencyArrowSeparator(props: CurrencyArrowSeparatorProps): Reac
       data-isLoading={isLoading ? true : undefined}
       isCollapsed={isCollapsed}
       hasSeparatorLine={hasSeparatorLine}
-      disabled={disabled}
     >
-      <styledEl.LoadingWrapper isLoading={isLoading} onClick={disabled ? undefined : onSwitchTokens}>
+      <styledEl.LoadingWrapper type="button" $isLoading={isLoading} disabled={disabled} onClick={onSwitchTokens}>
         {!isInjectedWidgetMode && isLoading ? (
           <CowLoadingIcon size={26} isDarkMode={isDarkMode} />
         ) : (