+++ libs/ui/CHANGELOG.md
@@ -1,5 +1,31 @@
 # Changelog
 
+## [2.2.2](https://github.com/cowprotocol/cowswap/compare/ui-v2.2.1...ui-v2.2.2) (2026-02-17)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @cowprotocol/analytics bumped to 2.2.2
+    * @cowprotocol/common-const bumped to 2.2.2
+    * @cowprotocol/common-hooks bumped to 2.2.2
+    * @cowprotocol/common-utils bumped to 2.2.2
+    * @cowprotocol/core bumped to 2.2.2
+
+## [2.2.1](https://github.com/cowprotocol/cowswap/compare/ui-v2.2.0...ui-v2.2.1) (2026-02-16)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @cowprotocol/analytics bumped to 2.2.1
+    * @cowprotocol/common-const bumped to 2.2.1
+    * @cowprotocol/common-hooks bumped to 2.2.1
+    * @cowprotocol/common-utils bumped to 2.2.1
+    * @cowprotocol/core bumped to 2.2.1
+
 ## [2.2.0](https://github.com/cowprotocol/cowswap/compare/ui-v2.1.0...ui-v2.2.0) (2026-02-13)